requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Langfuse
langfuse==2.20.0

# Google AI
google-genai==1.24.0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
httpx  # バージョン指定を削除して依存関係の解決をpipに任せる
json5==0.9.14
jinja2==3.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2